refactor: card를 별도 component로 분리 - Tinder 형식 인터페이스 제거. - 디자인 이슈 해결까지 개발 보류.
